+++ styleguide.config.js
@@ -108,6 +108,7 @@ module.exports = {
             'react/components/Button/index.js',
             'react/components/ButtonWithIcon/index.js',
             'react/components/Checkbox/index.js',
+            'react/components/ColorPicker/index.js',
             'react/components/DatePicker/index.js',
             'react/components/Dropdown/index.js',
             'react/components/ActionMenu/index.js',