+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+
+- **ColorPicker** component.
 
 ## [8.17.8] - 2019-02-11
 
@@ -104,7 +107,6 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - **`Toast`** Add `horizontalPosition` to `showToast` options.
 
 ## [8.12.0] - 2019-01-17
-
 ### Added
 
 - **Dropdown** `variation` prop